requirements.txt
# Core
fastapi==0.110.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# Integrations
stripe==7.12.0
openai==1.10.0
httpx==0.26.0

# Redis & Background Tasks
redis==5.0.1
celery==5.3.4

# Utils
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Monitoring
sentry-sdk[fastapi]==1.40.0
psutil==5.9.8

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
